[Refactoring] Seperate define value
